Move test and delegate port selection to OS
